Add an $added array to Translator to store added lines
Adding lines to the Translator prevents the Translator from loading the translation files, as it thinks it has already loaded. By storing the added lines in a separate array, and merging the 'loaded array' and the 'added array' when getting a line, you can prevent that behavior.

Steps to reproduce:
Set a different language (eg. nl) as main locale, and English as fallback.
Duplicate the lang/en/passwords.php file to the new language dir and update the values to something different.
In your view, echo
```
{{ __('passwords.reset') }}
{{ __('passwords.sent') }}
```

It should show your new language output.

If you add a line to the translator before the files are loaded (eg. in a service provider):

```
    /**
     * Bootstrap any application services.
     *
     * @return void
     */
    public function boot()
    {
        app('translator')->addLines(['passwords.reset' => 'UPDATED LINE'], 'nl');
    }
```

Now, if you open your view again, it shows the UPDATED LINE for the 'passwords.reset' entry, but the 'passwords.sent' entry shows the English fallback line.
This is because after adding the line to the nl locale, the Translator->isLoaded returns true for the current locale, as it is set by adding the line.

This update will fix that, as it stores added lines in a separate array.